requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
python-dotenv==1.0.1
virtualenv==20.26.2
django-storages==1.14.4